Added a temp fix for FontSize issue.
This temporary fix provides a Length-based constructor for Rules.FontSize() instead of the original. The original still exists and will exist until such time as Unity Technologies clarifies which USS data-type is to be provided for this style rule.

The issue is known to affect at least documentation for Unity Engine 2021.3 to 2023.2 (beta).