Updated document for v1.0.0-prerelease-2
The documentation introduces some new fixes, most notably addressing issues #1, #8, #9 , #10.

Renaming `USS.StyleSheet` to `USS.Sheet` was not a change I desired to make however due to `UnityEngine.UIElements.StyleSheet` it was a conflict that needed a swift name change for the type. Albeit undescriptive, this should solve the conflicts between the two. Please raise any issues up if any other conflicts are found.

Renaming `USS.Length` to `USS.Len` was another side effect of the **UIElements** namespace, though this easier to resolve.